+++ gaphor/core/styling/tests/test_css.py
@@ -1,6 +1,10 @@
 import pytest
 
-from gaphor.core.styling import CompiledStyleSheet, compile_style_sheet
+from gaphor.core.styling import (
+    CompiledStyleSheet,
+    compile_style_sheet,
+    compute_pseudo_element_style,
+)
 from gaphor.core.styling.declarations import WhiteSpace
 from gaphor.core.styling.tests.test_compiler import Node
 
@@ -151,7 +155,7 @@ def test_compiled_style_sheet():
     """
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
     assert props.get("font-family") == "sans"
     assert props.get("font-size") == 42
@@ -175,20 +179,24 @@ def test_faulty_font_size(font_size):
     assert props.get("font-size") is None
 
 
+def filter_private(props):
+    return {n: v for n, v in props.items() if not n.startswith("-gaphor-")}
+
+
 def test_empty_compiled_style_sheet():
     css = ""
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
-    assert props == {}
+    assert filter_private(props) == {}
 
 
 def test_color():
     css = "mytype { color: #00ff00 }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
     assert props.get("color") == (0, 1, 0, 1)
 
@@ -197,7 +205,7 @@ def test_color_typing_in_progress():
     css = "mytype { color: # }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
     assert props.get("color") is None
 
@@ -218,7 +226,7 @@ def test_line_style(css_value, result):
     css = f"mytype {{ line-style: {css_value} }}"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
     assert props.get("line-style") == result
 
@@ -238,7 +246,7 @@ def test_opacity(css_value, result):
     css = f"mytype {{ opacity: {css_value} }}"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
     assert props.get("opacity") == result
 
@@ -248,7 +256,7 @@ def test_broken_line_style():
     css = "diagram { line-style: sloppy * { }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("mytype"))
+    props = compiled_style_sheet.compute_style(Node("mytype"))
 
     assert props.get("line-style") is None
 
@@ -257,7 +265,7 @@ def test_variable():
     css = "diagram { --myvar: 12; line-width: var(--myvar) }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("line-width") == 12
 
@@ -266,7 +274,7 @@ def test_variable_color():
     css = "* { --mycolor: #123456 } diagram { color: var(--mycolor) }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("color") == pytest.approx(
         (0x12 / 255.0, 0x34 / 255.0, 0x56 / 255.0, 1.0)
@@ -277,7 +285,7 @@ def test_unknown_variable():
     css = "diagram { line-width: var(--myvar) }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("line-width") is None
 
@@ -286,7 +294,7 @@ def test_unknown_variable_should_use_original_value():
     css = "* { line-width: 1.0 } diagram { line-width: var(--myvar) }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("line-width") == pytest.approx(1.0)
 
@@ -295,7 +303,7 @@ def test_unknown_variable_resolve_original_value():
     css = "* { line-width: var(--lw); --lw: 1.0 } diagram { line-width: var(--myvar) }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("line-width") == pytest.approx(1.0)
 
@@ -304,7 +312,7 @@ def test_variable_cannot_contain_a_variable():
     css = "* { line-width: var(--a); --a: var(--b); --b: 1.0 }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("line-width") is None
 
@@ -313,7 +321,7 @@ def test_variable_with_property():
     css = "* { line-width: var(line-width); }"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("diagram"))
+    props = compiled_style_sheet.compute_style(Node("diagram"))
 
     assert props.get("line-width") is None
 
@@ -330,9 +338,9 @@ def test_color_schemes():
     """
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    normal_props = compiled_style_sheet.match(Node("node"))
-    dark_props = compiled_style_sheet.match(Node("node", dark_mode=True))
-    light_props = compiled_style_sheet.match(Node("node", dark_mode=False))
+    normal_props = compiled_style_sheet.compute_style(Node("node"))
+    dark_props = compiled_style_sheet.compute_style(Node("node", dark_mode=True))
+    light_props = compiled_style_sheet.compute_style(Node("node", dark_mode=False))
 
     assert normal_props.get("line-width") == pytest.approx(1.0)
     assert dark_props.get("line-width") == pytest.approx(2.0)
@@ -351,6 +359,46 @@ def test_white_space_normal(white_space_value, result):
     css = f"* {{ white-space: {white_space_value} }}"
 
     compiled_style_sheet = CompiledStyleSheet(css)
-    props = compiled_style_sheet.match(Node("node"))
+    props = compiled_style_sheet.compute_style(Node("node"))
 
     assert props.get("white-space") is result
+
+
+def test_pseudo_element():
+    css = "class::after { content: 'Hi' }"
+
+    compiled_style_sheet = CompiledStyleSheet(css)
+    props = compiled_style_sheet.compute_style(Node("class", pseudo="after"))
+
+    assert props.get("content") == "Hi"
+
+
+def test_bare_pseudo_element():
+    css = "::after { content: 'Hi' }"
+
+    compiled_style_sheet = CompiledStyleSheet(css)
+    props = compiled_style_sheet.compute_style(Node("node", pseudo="after"))
+
+    assert props.get("content") == "Hi"
+
+
+def test_combined_normal_and_pseudo_element():
+    css = "node, class::after { content: 'Hi' }"
+
+    compiled_style_sheet = CompiledStyleSheet(css)
+    node_props = compiled_style_sheet.compute_style(Node("node"))
+    class_props = compiled_style_sheet.compute_style(Node("class", pseudo="after"))
+
+    assert node_props.get("content") == "Hi"
+    assert class_props.get("content") == "Hi"
+
+
+def test_pseudo_element_inherits_normal_declarations():
+    css = "node { color: blue} ::after { content: 'Hi' }"
+
+    compiled_style_sheet = CompiledStyleSheet(css)
+    props = compiled_style_sheet.compute_style(Node("node"))
+    after = compute_pseudo_element_style(props, "after")
+
+    assert after
+    assert after.get("content") == "Hi"